requirements.txt
pandas==1.0.5
numpy==1.19.0
tensorflow==2.2.0
scikit_image==0.17.2
streamlit==0.69.2
opencv_contrib_python==4.4.0.44
matplotlib==3.3.0
Pillow==8.0.1
skimage==0.0
